requirements.txt
# Core dependencies for the web application
streamlit>=1.28.0
openai>=1.3.0
git+https://github.com/openai/whisper.git
python-docx>=0.8.11
requests>=2.31.0
beautifulsoup4>=4.12.0
mutagen>=1.46.0
tiktoken>=0.5.0

# Audio processing
ffmpeg-python>=0.2.0

# Optional: for better performance
torch>=2.0.0
torchaudio>=2.0.0